+++ src/main/java/frc/robot/Robot.java
@@ -14,6 +14,8 @@ import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import edu.wpi.first.wpilibj.SpeedControllerGroup;
 import edu.wpi.first.wpilibj.SpeedController;
 import edu.wpi.first.wpilibj.Timer;
+import frc.robot.sensors.ColorSensor;
+import frc.robot.libs.Xbox;
 
 /**
  * This is a demo program showing the use of the DifferentialDrive class.
