+++ src/main/java/frc/robot/dashboard/Dashboard.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.dashboard;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
