+++ src/main/java/frc/robot/libs/ControllerConstants.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.libs;
 
 public class ControllerConstants{
